requirements.txt
langchain>=0.1.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6
tiktoken>=0.5.2
streamlit>=1.32.0
requests>=2.31.0
PyPDF2>=3.0.0
python-docx>=1.0.1
unstructured>=0.12.0